Create Cam class in vision.py
Create a Cam class that inherits from the ServoMotor class. Have it be
able to look up and down at it's safest predefined angles (pulse
levels). Be conservative with these angles as camera is not attached
very securely to servo and there is the risk that it will fall off.