Restructure the menu and write line follow draft
Restructure the menu not to use so many elif statements and to
automatically align colons. Use a dict of tuples to hold all the actions
for each option and divide menu if statement based on how many arguments
each functionin the tuple takes. Also instead of having 'space' appear
in the menu just indicate with ' '.

Introducing line following concept. Write functions for the logic of
each possible scenario of interrupting the IR sensors. Have position and
motion dicts but don't actually use them at this point. Plan is to keep
track of the robot to see where it needs to turn based on if it has
passed the line and is trying to correct itself or not. Turning still
needs work. Write while loop to cause the robot to check IR sensors
every 0.1 seconds. Add option to menu to enter line following mode.

Temporarily set logging level to INFO as checking every 0.1 seconds
causes many logging messages to appear.