Add command history tracking
Remove redundant width and menu_str assignments in manual_mode(), taken
care of in generate_menu_str().

Use start_time and end_time to record the time each command takes for
being able to reproduce a series of commands in the exact order and
length (time wise) that a user inputs.