Account for incorrect commands input by user
Move code for adding commands to history list to beneath the if
statement checking if the command is even valid.

Assign previous_user_command before entering while loop to deal the
special case of the first command.

Move previous_user_command = user_command from before the input()
statement to the end of the loop. The reason for this is to avoid
assigning an invalid user entered command to previous_user_command. It
only executes when the user enters a correct command.

Add comments to describe the history updating process.

Add a robot.car.brake() call to the sensor case to prevent the robot
from continuing to drive while checking sensors.