Make the commands a Carm attribute
Make cmd_dict an attribute of Carm class instead of defining in
main.py. Get rid of the cmd_order and instead use a list of tupples
named cmd_list. Use list of tupples because may add more commands to the
list (e.g. "line") but want to keep the commands themselves immutable.

Restructure main.py to account for all of this.